% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_chen_ts.R
\name{arma_chen_ts}
\alias{arma_chen_ts}
\title{CHARMA fitting without regressors}
\usage{
arma_chen_ts(y, ar = NULL, ma = NULL, tau = 0.5)
}
\arguments{
\item{y}{The vector of variables to fit}

\item{ar}{Specified in description.}

\item{ma}{Specified in description.}

\item{tau}{The quantile}
}
\description{
Used for fitting the Chen auto-regressive moving averages model with regressors. The model is given by
\deqn{\beta_{0} +
\sum_{j\in ar}\phi_j \log(y_{t-j})
+\sum_{j\in ma}\theta_jr_{t-j}}
Where \itemize{
\item{\eqn{y} are the variables}
\item{\eqn{\beta_{0}} is the intercept}
\item{\eqn{\boldsymbol{x}_t} are the covariables}
\item{\eqn{\boldsymbol{\beta}} are the regression coefficients}
\item{\eqn{ar} are the indices for the auto-regression}
\item{\eqn{ma} are the indices for the moving-averages}
\item{\eqn{\phi} are the auto-regression coefficients}
\item{\eqn{\theta} are the moving-averages coefficients}
\item{\eqn{r} are the errors}}
}
