% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_chen.R
\name{cdf_chen}
\alias{cdf_chen}
\title{The Cumulative Distribution Function of the Chen Distribution}
\usage{
cdf_chen(y, theta)
}
\arguments{
\item{y}{A numeric. The input of the mathematical function}

\item{theta}{A length 2 numeric vector with the parameters.}
}
\value{
A double representing the cumulative probability till y \code{y}
}
\description{
The mathematical function that describes the cumulative probability for the Chen
 distribution. Its formula is \deqn{F(y| \lambda,\delta)= 1 - \exp \left[  \delta \left( 1-\exp(y^{\lambda}) \right) \right].}.
}
\examples{
#same numbers as the cdf_chen function
curve(cdf_chen(x, theta = c(0.4, 0.01)), from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(0.4, 0.1)), from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(0.7, 0.01)), from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(1, 0.01)), from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(1, 0.0001)), from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
}
