% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_chen_rpr.R
\name{cdf_chen_rpr}
\alias{cdf_chen_rpr}
\title{The Cumulative Distribution Function of the Reparameterized Chen Distribution}
\usage{
cdf_chen_rpr(y, theta, tau = 0.5, lower_tail = T)
}
\arguments{
\item{y}{A numeric. The input of the mathematical function}

\item{theta}{A length 2 numeric vector, or coercible to vector using `as.vector(unlist())`,
with the parameters.}

\item{tau}{A single number between 0 and 1 that is the quantile.}
}
\value{
A double representing the cumulative probability till \code{y}
}
\description{
The mathematical function that describes the cumulative distribution for the Reparameterized
Chen distribution. Its formula is \deqn{F(y | \lambda, \mu, \tau)= 1 - \exp
\left[ \dfrac{\log(1-\tau)}{(1-\exp(\mu^{\lambda}))} \left(  1-\exp\left(y^{\lambda}\right)
\right) \right] , \quad  y > 0.}
}
\examples{
# same numbers as the cdf_chen_rpr function
curve(cdf_chen_rpr(x, theta = c(0.7, 7), tau = 0.5),
      from = 0, to = 20, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen_rpr(x, theta = c(0.5, 2), tau = 0.3),
      from = 0, to = 20, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen_rpr(x, theta = c(0.9, 3)),
      from = 0, to = 20, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen_rpr(x, theta = c(0.7, 11)),
      from = 0, to = 20, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen_rpr(x, theta = c(0.2, 8), tau = 0.2),
      from = 0, to = 20, xlab = "y", ylab = "cdf", n = 1000)
}
