% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_chen.R
\name{estim_chen}
\alias{estim_chen}
\title{Parameter estimation for the reparameterized Chen distribution}
\usage{
estim_chen(y, clvl = NULL, n_bootstrap = NULL)
}
\arguments{
\item{y}{A numeric vector with the random sample}

\item{clvl}{The confidence level for calculating the confidence intervals. If NULL no confidence
intervals will be calculated. Default in NULL.}

\item{n_bootstrap}{The number of resamples for bootstrap correction. Default is NULL,
so no boostrap correction occurs.}
}
\value{
If clvl is set to NULL (default) and full is set to TRUE(default)
 than the return is lenght 2 numeric vector with the parameter estimations. If
 clvl is a number, than the return is a 3x2 matrix with the parameter estimations
 and their confidence intervals. If full is set to TRUE than the return is a
 list, having by default 6 elements, if clvl is a number than the list will
 have 7 elements with the last one being the matrice of confidence intervals.
}
\description{
The Maximum likelihood estimate for the pair of parameters is numerically
approximated with \code{stats::optim} from the stats package.
}
\examples{
estim_chen(rchen(100, c(1,1)))
estim_chen(rchen(100, c(1,1)), clvl = 0.95)
estim_chen(rchen(100, c(1,1)), clvl = 0.95, n_bootstrap = 100)

}
