% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_chen_rpr.R
\name{estim_chen_rpr}
\alias{estim_chen_rpr}
\title{Parameter estimation for the reparameterized Chen distribution}
\usage{
estim_chen_rpr(y, method = "BFGS", tau = 0.5, full = F, ci_alpha = NULL)
}
\arguments{
\item{y}{Values sampled from a reparameterized Chen distribution.}

\item{method}{The optimisation method useed by optim, default is BFGS}

\item{tau}{The quantile for the reparameterized distribution}

\item{full}{If the return should be the entire list by optim or just the parameter
estimations}

\item{ci_alpha}{The alpha for calculating the confidence intervals. If null no confidence
intervals will be calculated. Default in NULL}
}
\value{
The pair of estimated parameters. If full == TRUE than will return the same list as the optim function.
If CI = TRUE will return the confidence interval.
}
\description{
The Maximum likelihood estimate for the pair of parameters is numerically approximated with optim from the stats package. The estim_chen function is effectivelly a wrapper for optim, but it defines and uses the Chen log-likelihood function as well.
}
\examples{
estim_chen_rpr(rchen(100, c(0.7,7)), tau = 0.5)

}
