% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_vec}
\alias{extend_vec}
\title{Extend vector}
\usage{
extend_vec(vec, n)
}
\arguments{
\item{vec}{The original vector}

\item{n}{The desired length}
}
\value{
Extended vector
}
\description{
Extends a vecotr by adding zeros to the end until the desired length is reached, the desired
length needs to be more than the original length.
}
