% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null}
\alias{is_null}
\title{Check if a vector is NULL.}
\usage{
is_null(vec)
}
\arguments{
\item{vec}{The vector to check.}
}
\value{
TRUE if the vector is NULL, FALSE otherwise.
}
\description{
This function checks if the provided vector is identical to NULL.
}
\keyword{internal}
