% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_chen.R
\name{ll_chen}
\alias{ll_chen}
\title{Log-likelihood function for the Chen distribution}
\usage{
ll_chen(y, theta)
}
\arguments{
\item{theta}{}
}
\description{
Mathematical formula given by \deqn{
\ell(\theta | y) = \log(\log(1 - \tau)) - \log(1 - \exp(\mu_t^{\lambda})) +
(\lambda - 1) \log(y_t) + \log(\lambda) + \frac{\log(1 - \tau)(1 - \exp(y_t^{\lambda}))}
{1 - \exp(\mu_t^{\lambda})} + y_t^{\lambda}
}
Where \eqn{\mu_t} and \eqn{\lambda} are parameters, \eqn{\tau} is quantile
and \eqn{y} is the random sample.
}
