% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_chen_ts_utils.R
\name{ll_REG_AR}
\alias{ll_REG_AR}
\title{Title}
\usage{
ll_REG_AR(y, y_cut, log_y, theta, n, n_ar, ar, max_ar, cvar, tau)

ll_REG_AR(y, y_cut, log_y, theta, n, n_ar, ar, max_ar, cvar, tau)
}
\arguments{
\item{cvar}{}
}
\description{
Title

Title
}
\keyword{internal}
