% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_chen_ts_utils.R
\name{ll_REG_ARMA}
\alias{ll_REG_ARMA}
\title{Title}
\usage{
ll_REG_ARMA(y, y_cut, log_y, theta, n, n_ar, n_ma, ar, ma, max_arma, cvar, tau)
}
\arguments{
\item{max_arma}{}
}
\description{
Title
}
\keyword{internal}
