% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.reg_chen}
\alias{plot.reg_chen}
\alias{plot.CHARMA}
\alias{plot.reg_CHARMA}
\title{Plots for reg_chen, CHARMA and reg_ARMA objects}
\usage{
\method{plot}{reg_chen}(x, layout = "all")

\method{plot}{CHARMA}(x, layout = "all")

\method{plot}{reg_CHARMA}(x, layout = "all")
}
\arguments{
\item{x}{The object}

\item{layout}{Could be \code{"all"} so all plots are plotted in the same image (2x2 grid)
or \code{"sep"} so plots are displayed separatelly.}
}
\value{
The plotted graphs
}
\description{
Plots for reg_chen, CHARMA and reg_ARMA objects
}
\section{Functions}{
\itemize{
\item \code{plot(CHARMA)}: CHARMA objects

\item \code{plot(reg_CHARMA)}: reg_CHARMA objects

}}
