% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.CHARMA}
\alias{print.CHARMA}
\title{Printing for CHARMA objects}
\usage{
\method{print}{CHARMA}(x, ...)
}
\arguments{
\item{x}{CHARMA object}

\item{...}{Further arguments to be passedo or from other methods}
}
\description{
Printing for CHARMA objects
}
