% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.reg_CHARMA}
\alias{print.reg_CHARMA}
\title{Printing for reg_CHARMA objects}
\usage{
\method{print}{reg_CHARMA}(x, ...)
}
\arguments{
\item{x}{reg_CHARMA object}

\item{...}{Further arguments to be passedo or from other methods}
}
\description{
Printing for reg_CHARMA objects
}
