% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.reg_chen}
\alias{print.reg_chen}
\title{Printing for reg_chen objects}
\usage{
\method{print}{reg_chen}(x, ...)
}
\arguments{
\item{x}{reg_chen object}

\item{...}{Further arguments to be passedo or from other methods}
}
\description{
Printing for reg_chen objects
}
