% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_chen_ts_utils.R
\name{reg_arma_case}
\alias{reg_arma_case}
\title{Determines the ARMA case}
\usage{
reg_arma_case(isar, isma, isreg)
}
\arguments{
\item{isar}{A logical indicating if the model incorporates auto regression}

\item{isma}{A logical indicating if the model incorporates moving averages}

\item{isreg}{A logical indicating if the model incorporates regular regression
(with covariables)}
}
\value{
A string with the ARMA case
}
\description{
Simply that determines the ARMA case between ARMA, AR, MA, REG_ARMA, REG_MA and
REG_AR. This is an internal function.
}
\examples{
# No need for examples in internal function

}
\keyword{internal}
