% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_chen.R
\name{reg_chen}
\alias{reg_chen}
\title{Title}
\usage{
reg_chen(data, formula, tau = 0.5, n_bootstrap = NULL)
}
\arguments{
\item{data}{A data.frame(or coercible to data.frame)}

\item{formula}{The formula for the regression}

\item{tau}{The quantile}

\item{n_bootstrap}{The number of resamples for bootstrap correction. Default is NULL,
so no boostrap correction occurs.}
}
\value{
The linear model to predict the values of mu
}
\description{
Title
}
