% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funcs.R
\name{reg_chen_predict}
\alias{reg_chen_predict}
\title{Prediction for the Chen regression model}
\usage{
reg_chen_predict(x, data, formula = NULL)
}
\arguments{
\item{x}{An object fitted by \code{reg_chen}}

\item{data}{A matrix with the predictors.}

\item{formula}{A \code{formula} object. This is an optional parameter, if no formula is provided
than \code{data} is considered to be the already formated matrix of covariables. It is recommended
to send a \code{formula} object to avoid errors.}
}
\value{
The vector of predictions.
}
\description{
This function is used for predictions using regression models fitted by \code{reg_chen}
}
