% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_chen.R
\name{cdf_chen}
\alias{cdf_chen}
\title{The Cumulative Distribution Function of the Chen Distribution}
\usage{
cdf_chen(y, theta, lower_tail = T)
}
\arguments{
\item{y}{The input of the mathematical function, if it is a vector, than the return will be a vector.}

\item{theta}{A length 2 numeric vector, or coercible to vector using \code{as.vector(unlist(theta))},
with the parameters.}

\item{lower_tail}{If true, than probabilities are \eqn{P(Y \leq y)}, otherwise \eqn{P(Y \geq y)}}
}
\value{
A double, or vector of doubles, representing the cumulative probability till \code{y}
}
\description{
The mathematical function that describes the cumulative probability for the Chen
 distribution. Its formula is \deqn{F(y| \lambda,\delta)= 1 - \exp \left[  \delta \left( 1-\exp(y^{\lambda}) \right) \right]}
}
\examples{
# same numbers as the pdf_chen function
cdf_chen(1, theta = c(0.7, 0.2))
curve(cdf_chen(x, theta = c(0.4, 0.1)),
      from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(0.7, 0.01)),
      from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(1, 0.01)),
      from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
curve(cdf_chen(x, theta = c(1, 0.0001)),
      from = 0, to = 100, xlab = "y", ylab = "cdf", n = 1000)
}
