% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_chen.R
\name{estim_chen}
\alias{estim_chen}
\title{Parameter estimation for the Chen distribution}
\usage{
estim_chen(y, method = "BFGS")
}
\arguments{
\item{y}{Values sampled from a Chen distribution.}
}
\value{
The best set of parameters found
}
\description{
The Maximum likelihood estimate for the pair of parameters is numerically approximated with optim from the stats package. The estim_chen function is effectivelly a wrapper for optim, but it defines and uses the Chen log-likelihood function as well.
}
\examples{
estim_chen(rchen(10, c(1,1)))

}
