% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_chen.R
\name{estim_chen}
\alias{estim_chen}
\title{Parameter estimation for the reparameterized Chen distribution}
\usage{
estim_chen(y, method = "BFGS", full = F, clvl = NULL)
}
\arguments{
\item{y}{Values sampled from a reparameterized Chen distribution.}

\item{method}{The optimisation method used by optim, default is BFGS}

\item{full}{If the return should be the entire list by optim or just the parameter
estimations}

\item{clvl}{The confidence level for calculating the confidence intervals. If NULL no confidence
intervals will be calculated. Default in NULL.}
}
\value{
If clvl is set to NULL (default) and full is set to TRUE(default)
 than the return is lenght 2 numeric vector with the parameter estimations. If
 clvl is a number, than the return is a 3x2 matrix with the parameter estimations
 and their confidence intervals. If full is set to TRUE than the return is a
 list, having by default 6 elements, if clvl is a number than the list will
 have 7 elements with the last one being the matrice of confidence intervals.
}
\description{
The Maximum likelihood estimate for the pair of parameters is numerically
approximated with optim from the stats package. The estim_chen function is
effectivelly a wrapper for optim, but it defines and uses the Chen
log-likelihood function as well.
}
\examples{
estim_chen(rchen(10, c(1,1)))

}
