% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_chen_rpr.R
\name{estim_chen_rpr}
\alias{estim_chen_rpr}
\title{Parameter estimation for the reparameterized Chen distribution}
\usage{
estim_chen_rpr(y, tau = 0.5, clvl = NULL, n_bootstrap = NULL)
}
\arguments{
\item{y}{A numeric vector, or coercible to vector using `as.vector(unlist())`.
Values sampled from a reparameterized Chen distribution.}

\item{tau}{The quantile for the reparameterized distribution}

\item{clvl}{The confidence level for calculating the confidence intervals. If NULL no confidence
intervals will be calculated. Default in NULL.}

\item{n_bootstrap}{The number of resamples for bootstrap correction. Default is NULL,
so no boostrap correction occurs.}
}
\value{
If clvl is set to NULL (default) and full is set to TRUE(default)
 than the return is lenght 2 numeric vector with the parameter estimations. If
 clvl is a number, than the return is a 3x2 matrix with the parameter estimations
 and their confidence intervals. If full is set to TRUE than the return is a
 list, having by default 6 elements, if clvl is a number than the list will
 have 7 elements with the last one being the matrice of confidence intervals.
}
\description{
The Maximum likelihood estimate for the pair of parameters is numerically
approximated with optim from the stats package. The estim_chen_rpr function is
effectivelly a wrapper for optim, but it defines and uses the Chen
log-likelihood function as well.
}
\examples{
estim_chen_rpr(rchen(100, c(0.7,7)), tau = 0.5)
estim_chen_rpr(rchen(100, c(0.7,7)), tau = 0.5, clvl = 0.95)
estim_chen_rpr(rchen(100, c(0.7,7)), tau = 0.5, clvl = 0.95, n_bootstrap = 100)

}
