% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null}
\alias{is_null}
\title{Check if a vector is NULL.}
\usage{
is_null(vec)
}
\arguments{
\item{vec}{The vector to check.}
}
\value{
TRUE if the vector is NULL, FALSE otherwise.
}
\description{
This function checks if the provided vector is identical to NULL. This was created
since I found  no way to, in base R, check if an entire vector is NULL. We would usually need to
use \code{any(is.null(vec))}. I differed the syntax from base R's is.null by using an underscore
instead of a . (\code{is_null} instead of \code{is.null})
}
\examples{
# No need for examples in internal function

}
\keyword{internal}
