% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_chen.R
\name{pdf_chen}
\alias{pdf_chen}
\title{The Probability Density Function of the Chen Distribution}
\usage{
pdf_chen(y, theta)
}
\arguments{
\item{y}{A numeric. The input of the mathematical function}

\item{theta}{A length 2 numeric vector with the parameters.}
}
\value{
A double representing the probability density at \code{y}
}
\description{
The mathematical function that describes the probability density for the Chen
 distribution. Its formula is \deqn{f(y|\lambda, \delta)= \delta\lambda
 y^{\lambda - 1} \exp \left\lbrace \delta \left[ 1-\exp(y^{\lambda})\right]
 +y^{\lambda} \right\rbrace  , \quad y>0.}
}
\examples{
#same numbers as the pdf_chen function
curve(pdf_chen(x, theta = c(0.4, 0.01)), from = 0, to = 100, xlab = "y", ylab = "pdf", n = 1000)
curve(pdf_chen(x, theta = c(0.4, 0.1)), from = 0, to = 100, xlab = "y", ylab = "pdf", n = 1000)
curve(pdf_chen(x, theta = c(0.7, 0.01)), from = 0, to = 100, xlab = "y", ylab = "pdf", n = 1000)
curve(pdf_chen(x, theta = c(1, 0.01)), from = 0, to = 100, xlab = "y", ylab = "pdf", n = 1000)
curve(pdf_chen(x, theta = c(1, 0.0001)), from = 0, to = 100, xlab = "y", ylab = "pdf", n = 1000)
}
