% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_chen.R
\name{pdf_chen}
\alias{pdf_chen}
\title{The Chen Probability Density Function}
\usage{
pdf_chen(y, theta)
}
\arguments{
\item{y}{A numeric. The input of the mathematical function}

\item{theta}{A length 2 numeric vector with the parameters.}
}
\value{
A double representing the probability density at \code{y}
}
\description{
The mathematical function that describes the probability density for the Chen
 distribution. Its formula is \deqn{f(y|\lambda, \delta)= \delta\lambda
 y^{\lambda - 1} \exp \left\lbrace \delta \left[ 1-\exp(y^{\lambda})\right]
 +y^{\lambda} \right\rbrace  , \quad y>0.}Usually used for plotting.
}
\examples{
curve(pdf_chen(x, c(0.7, 0.1)), from = 0, to = 10, xlab = "y",
  ylab = "Probability density")
}
