% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchen.R
\name{rchen}
\alias{rchen}
\title{Random generation for the chen distribution with parameters lambda and delta}
\usage{
rchen(n, theta)
}
\arguments{
\item{n}{number of random values to generate}

\item{theta}{A length 2 numeric vector. The first element is the parameter lambda, and the second one is delta, both need to be bigger than 0.}
}
\value{
A length n numeric vector.
}
\description{
Random generation for the chen distribution with parameters lambda and delta
}
\examples{
rchen(c(0.7, 0.4), 30)
rchen(c(0.1, 0.1), 10)
rchen(c(1, 0.3), 10)
rchen(c(0.7, 0.01), 10)

}
