% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchen_rpr.R
\name{rchen_rpr}
\alias{rchen_rpr}
\title{Random generation for the Reparameterized Chen distribution}
\usage{
rchen_rpr(n, theta, tau = 0.5)
}
\arguments{
\item{n}{Number of random values to generate}

\item{theta}{A length 2 or n+1 vector or list with lambda and mu respectively.
The values will be coersed to vector using \code{as.vector(unlist(theta))} and
lambda will always be considered as the first element in that vector, if the length is two than all values will
be generated with the same mu, if the length is n+1 than one value will be generated with each mu.}

\item{tau}{The quantile.}
}
\value{
A length n numeric vector with the random values.
}
\description{
Used for generating random variables. Specially usefull for simulation
purposes.
}
\examples{
rchen_rpr(10, c(0.7, 7), 0.5)
rchen_rpr(10, c(0.1, 1))
rchen_rpr(100, list(0.7, stats::runif(100, 6,7)), .3)
hist(rchen_rpr(100, c(0.7, 7), 0.7))

}
