% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchen_ts.R
\name{rchen_ts}
\alias{rchen_ts}
\title{Random generation}
\usage{
rchen_ts(
  n,
  intercept,
  lambda,
  ar_coef = NULL,
  ma_coef = NULL,
  reg_coef = NULL,
  cvar = NULL,
  freq = 1
)
}
\arguments{
\item{n}{Number of random values}

\item{intercept}{The formula's intercept}

\item{lambda}{The distribution parameter lambda}

\item{ar_coef}{The coefficients for the auto-regressive model}

\item{ma_coef}{The coefficients for the moving averages model}

\item{freq}{Parameter for the time.series object that is returned}
}
\value{
A time serie object with the generated values
}
\description{
Random generation
}
\examples{
plot(rchen_ts(100, 1, 0.7, ar_coef = c(0.6, 0.1)))
plot(rchen_ts(100, 1, 0.7, ma_coef = c(0.5, 0.2)))
rchen_ts(100, 1, 0.7, ar_coef = c(0.5, 0.3), ma_coef = c(0.3, 0.1))

}
