% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_chen.R
\name{reg_chen}
\alias{reg_chen}
\title{Title}
\usage{
reg_chen(data, formula, tau = 0.5, stripped = F)
}
\arguments{
\item{data}{A data.frame(or coercible to data.frame)}

\item{formula}{The formula for the regression}

\item{tau}{The quantile}

\item{stripped}{Nothing yet...}
}
\value{
The linear model to predict the values of mu
}
\description{
Title
}
\examples{
a <- runif(100)
b <- runif(100)
form <- y ~ a + b
true_reg_coefficients <- c(2, 1)
mus <- exp(cbind(a, b) \%*\% matrix(true_reg_coefficients))
df <- data.frame(a = a, b = b, y = rchen(100, list(0.7, mus, 0.3)))
reg_chen(df, form, tau = 0.3)

}
