% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_chen_ts.R
\name{reg_chen_ts}
\alias{reg_chen_ts}
\title{CHARMA fitting with regressors}
\usage{
reg_chen_ts(data, formula, ar = NULL, ma = NULL, tau = 0.5)
}
\arguments{
\item{data}{A \code{data.frame} with the variables specified in \code{formula}}

\item{formula}{An object of class \code{formula} which is gonna be used for fitting
the model}

\item{ar}{Specified in description.}

\item{ma}{Specified in description.}

\item{tau}{The quantile}
}
\value{
An object of class \code{reg_CHARMA} with the following attributes:
\describe{

  \item{\code{coef}}{The coefficients of the regression model. This is what is gonna
  be used for predicting new data.}

  \item{\code{beta0}, \code{phi}, \code{theta}, \code{betas}}{Explained in description}

  \item{\code{lambda}}{The predicted value of the \eqn{\lambda} parameter in
  reparameterized Chen distribution formula.}

  \item{\code{hessian}}{A matrix giving the estimate of the Hessian at the solutions found}

  \item{\code{fitted}}{The output variables used for prediction, will be a column of \code{data}}

  \item{\code{etahat}}{The predicted variables in log scale}

  \item{\code{errorhat}}{The errors in prediction in log scale}

  \item{\code{case}}{The model case between, REG_ARMA, REG_AR and REG_MA, mainly used internally}}
}
\description{
Used for fitting the Chen auto-regressive moving averages model with regressors. The model is given by
\deqn{\eta_t = \beta_{0} + \boldsymbol{x}_t^\top\boldsymbol{\beta}+
\sum_{j\in ar}\phi_j [\log(y_{t-j})-\boldsymbol{x}_{t-j}^\top\boldsymbol{\beta}]
+\sum_{j\in ma}\theta_jr_{t-j}}
Where \itemize{
\item{\eqn{y} are the variables}
\item{\eqn{\beta_{0}} is the intercept}
\item{\eqn{\boldsymbol{x}} are the covariables}
\item{\eqn{\boldsymbol{\beta}} are the regression coefficients}
\item{\eqn{ar} are the indices for the auto-regression}
\item{\eqn{ma} are the indices for the moving-averages}
\item{\eqn{\phi} are the auto-regression coefficients}
\item{\eqn{\theta} are the moving-averages coefficients}
\item{\eqn{r} are the errors}}
}
\examples{
a <- stats::runif(100)
b <- stats::runif(100, 2, 3)
form <- y ~ a + b
df <- data.frame(a = a, b = b)
df$y <- chen::rchen_ts(100, 1, 0.7, ar_coef = c(0.2, 0.1), ma_coef = c(0.5, 0.3),
                                    reg_coef = c(0.2, 1), cvar = cbind(a, b))
reg_chen_ts(df, form, ar = 1:2, ma = 1:2)
df$y <- chen::rchen_ts(100, 1, 0.7, ar_coef = c(0.2, 0.1) , reg_coef = c(0.2, 1),
                                    cvar = cbind(a, b))

reg_chen_ts(df, form, ar = 1:2)
df$y <- chen::rchen_ts(100, 1, 0.7, ma_coef = c(0.5, 0.2), reg_coef = c(0.2, 1),
                                    cvar = cbind(a, b))
reg_chen_ts(df, form, ma = 1:2)
}
